🚀 : – Split arithmetic and secret helpers
- what: break out arithmetic and secret helpers into dedicated modules

- why: align with roadmap item for clearer separation

- how to test: pytest --cov=gabriel --cov-report=term-missing
